make require_match a propery of webjumps instead of completions
Nested completers present a hard problem with respect to the require_match
property, and where exactly this flag belongs when a nested completer,
like webjump_completer, needs it (in some cases).  As a UI concern, it is
clear that it breaks abstraction to use completions objects to carry this
flag from webjump to reader, and it necessitates overmuch plumbing to do
so.

It makes more sense to put the property on webjumps, because it is in a
very literal sense a property of the webjumps that a match is required.
But the reader doesn't know anything about webjumps, so in order to
communicate with the reader, we need to carry the flag from webjump to
reader via the completer.  This still needs explicit support from
merged_completer, and it is still a break of abstraction, but a lesser
one, and we eliminate fragile checks for the 'get_require_match' property
to dispatch special behavior.